+++ src/main/java/mod/beethoven92/betterendforge/common/world/feature/DesertLakeFeature.java
@@ -160,7 +160,7 @@ public class DesertLakeFeature extends Feature<NoFeatureConfig> {
 			}
 		}
 
-		double aspect = ((double) radius / (double) depth);
+		double aspect = (radius / depth);
 
 		for (int x = blockPos.getX() - dist; x <= blockPos.getX() + dist; x++) {
 			POS.setX(x);
